feat(cli): prepend atomic-commit discipline preamble to employee prompts (#717)
When csa run dispatches a prompt to an employee (codex, claude-code,
gemini-cli, opencode), prepend an <atomic-commit-discipline> preamble
that instructs the employee to commit after EACH logical change instead
of batching multiple unrelated changes into one commit.

Implements Option 1 (prompt-only) from #717. No new helper binary, no
skill changes. Over-prepending is the chosen trade-off: better a harmless
preamble in a no-commit context than a missed preamble in a fix dispatch.

Closes #717.
